Refactored the CommandInputFormatter to inherit from the JsonInputFormatter. I then also followed the model that the JsonInputFormatter gets added to the Service Collection by using an IConfigureOptions interface. Additionally I pulled the actual adding of the service descriptors into its own extension class and method to help clean up the Startup file.

+++ src/Vigil.WebApi/Controllers/PatronController.cs
@@ -58,7 +58,7 @@ namespace Vigil.WebApi.Controllers
         {
             using (var context = ContextFactory())
             {
-                if (context.Patrons.Any(p => p.Id == patronId))
+                if (context.Patrons.Any(p => p.Id == patronId && p.DeletedOn == null))
                 {
                     CommandQueue.Publish(new DeletePatron(User.Identity.Name ?? "Anonymous User", DateTime.UtcNow) { PatronId = patronId });
                     return Accepted();
